+++ examples/lang/js_web_api/main.rs
@@ -43,7 +43,7 @@ pub extern "C" fn main() {
     Js::eval_timeout("console.log('Delayed!');", 1000);
     Js::eval_interval("console.log('Repeating!');", 2000);
     let cleared = Js::eval_timeout("console.error('This should not run!');", 1500);
-    Js::eval_timeout_clear(cleared);
+    Js::clear_timeout(cleared);
 
     /* draw shapes */